feat(v2): Put together the constants of the models package
