--- v2/models/channeltype.go
@@ -1,14 +1,0 @@
-//
-// Copyright (C) 2020-2021 IOTech Ltd
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package models
-
-// ChannelType controls the range of values which constitute valid delivery types for channels
-type ChannelType string
-
-const (
-	Rest  = "REST"
-	Email = "EMAIL"
-)
