--- v2/models/notificationseverity.go
@@ -1,14 +1,0 @@
-//
-// Copyright (C) 2021 IOTech Ltd
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package models
-
-type NotificationSeverity string
-
-const (
-	Minor    = "MINOR"
-	Critical = "CRITICAL"
-	Normal   = "NORMAL"
-)
