--- v2/models/operatingstate.go
@@ -1,20 +1,0 @@
-//
-// Copyright (C) 2020 IOTech Ltd
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package models
-
-// OperatingState Constant String
-type OperatingState string
-
-/*
-	Up      : UP
-	Down    : DOWN
-	Unknown : UNKNOWN
-*/
-const (
-	Up      = "UP"
-	Down    = "DOWN"
-	Unknown = "UNKNOWN"
-)
