--- v2/models/transmissionstatus.go
@@ -1,15 +1,0 @@
-//
-// Copyright (C) 2021 IOTech Ltd
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package models
-
-type TransmissionStatus string
-
-const (
-	Failed       = "FAILED"
-	Sent         = "SENT"
-	Acknowledged = "ACKNOWLEDGED"
-	Trxescalated = "ESCALATED"
-)
